refactor: table MappedSuperClass 제거
