+++ src/main/java/com/debatetimer/dto/member/TableResponse.java
@@ -1,10 +1,10 @@
 package com.debatetimer.dto.member;
 
-import com.debatetimer.domain.DebateTable;
+import com.debatetimer.domain.customize.CustomizeTable;
 
 public record TableResponse(long id, String name, TableType type, String agenda) {
 
-    public TableResponse(DebateTable debateTable) {
+    public TableResponse(CustomizeTable debateTable) {
         this(
                 debateTable.getId(),
                 debateTable.getName(),
